# specs/001-review-comic-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PoSeeReview API
  description: |
    REST API for the PoSeeReview application - generates comic strips from strange restaurant reviews.
    
    **Key Features:**
    - Discover nearby restaurants using geolocation
    - Generate AI-powered comics from reviews
    - View real-time strangeness leaderboard
    
    **Tech Stack:** ASP.NET Core 9.0, Azure Table Storage, Azure OpenAI, DALL-E 3
  version: 1.0.0
  contact:
    name: PoSeeReview Team
  license:
    name: MIT
servers:
  - url: https://localhost:5001
    description: Local development (HTTPS)
  - url: http://localhost:5000
    description: Local development (HTTP)
  - url: https://poseereview.azurewebsites.net
    description: Production (Azure App Service)
...
paths:
  /api/restaurants/nearby:
    get:
      tags:
        - Restaurants
      summary: Get nearby restaurants
      description: |
        Discovers restaurants within 5km radius of provided coordinates using Google Maps API.
        Results are cached for 24 hours to minimize API costs.
      operationId: getNearbyRestaurants
      parameters:
        - name: latitude
          in: query
          required: true
          description: User's latitude (-90 to 90)
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          example: 47.6062
        - name: longitude
          in: query
          required: true
          description: User's longitude (-180 to 180)
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          example: -122.3321
        - name: limit
          in: query
          required: false
          description: Maximum number of results (default 10)
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Successfully retrieved restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurants:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestaurantDto'
                  totalCount:
                    type: integer
                    description: Total restaurants found
                  cachedAt:
                    type: string
                    format: date-time
                    description: When this result was cached
              example:
                restaurants:
                  - placeId: "ChIJabcdef123"
                    name: "The Flying Saucer Café"
                    address: "123 Space Needle Way, Seattle, WA"
                    latitude: 47.6205
                    longitude: -122.3493
                    averageRating: 4.5
                    totalReviews: 247
                    distance: 1.2
                totalCount: 8
                cachedAt: "2025-01-22T14:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/restaurants/{placeId}:
    get:
      tags:
        - Restaurants
      summary: Get restaurant details
      description: Retrieves detailed information about a specific restaurant, including reviews.
      operationId: getRestaurantDetails
      parameters:
        - name: placeId
          in: path
          required: true
          description: Google Maps place ID
          schema:
            type: string
          example: "ChIJabcdef123"
      responses:
        '200':
          description: Successfully retrieved restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantDetailsDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  schemas:
    RestaurantDto:
      type: object
      required:
        - placeId
        - name
        - address
        - latitude
        - longitude
      properties:
        placeId:
          type: string
          description: Google Maps place ID (unique identifier)
        name:
          type: string
          maxLength: 255
        address:
          type: string
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        averageRating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          description: Google average rating
        totalReviews:
          type: integer
          minimum: 0
        distance:
          type: number
          format: double
          description: Distance from user in kilometers
    RestaurantDetailsDto:
      allOf:
        - $ref: '#/components/schemas/RestaurantDto'
        - type: object
          properties:
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/ReviewDto'
              maxItems: 10
              description: Top 10 reviews sorted by strangeness score
    ReviewDto:
      type: object
      required:
        - text
        - rating
        - strangenessScore
      properties:
        authorName:
          type: string
        text:
          type: string
          minLength: 5
          description: Review content (minimum 5 words)
        rating:
          type: integer
          minimum: 1
          maximum: 5
        time:
          type: string
          format: date-time
        strangenessScore:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: AI-calculated strangeness (0=normal, 100=very strange)
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
        traceId:
          type: string
          description: Correlation ID for logging (W3C Trace Context format)
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://tools.ietf.org/html/rfc7231#section-6.5.1"
            title: "Bad Request"
            status: 400
            detail: "Invalid latitude: must be between -90 and 90"
            instance: "/api/restaurants/nearby"
            traceId: "00-abc123-def456-00"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://tools.ietf.org/html/rfc7231#section-6.5.4"
            title: "Not Found"
            status: 404
            detail: "Restaurant with placeId 'ChIJinvalid' does not exist"
            instance: "/api/restaurants/ChIJinvalid"
            traceId: "00-xyz789-uvw123-00"
    ServiceUnavailable:
      description: External dependency unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://tools.ietf.org/html/rfc7231#section-6.6.4"
            title: "Service Unavailable"
            status: 503
            detail: "Google Maps API is temporarily unavailable"
            instance: "/api/restaurants/nearby"
            traceId: "00-qrs456-tuv789-00"
  securitySchemes: {}
security: []